[wni] Refactor cli for other cloud providers
This commit refactors the existing cli to allow other cloud
provider implementations to reuse the base command. Added
a aws reference implementation
