+++ tools/windows-node-installer/main.go
@@ -1,0 +1,10 @@
+package main
+
+import (
+	"github.com/openshift/windows-machine-config-operator/tools/windows-node-installer/cmd"
+)
+
+// main is the entry point for windows-node-installer.
+func main() {
+	cmd.Execute()
+}
